extract permission commands into separate module
